Revise aarch64 kernel to use SIMD intrinsics. The structure is the same as the AVX 2 / FMA kernel, but the tile size is set to 8x8 as that performed best. I did not port the prefetching logic as `std::arch::aarch64::_prefetch` is currently nightly only :(

On an M1 Mac (AWS mac2.metal) performance for an M=N=K=1024 matmul increases from ~334 to ~418 GFLOPS.
On a Graviton 2 (AWS c6g.xlarge) performance for an M=N=K=1024 matmul increases from ~115 to ~136 GFLOPS.

One observation from my tests is that ARM performance is much more sensitive to whether unrolling is used than the AVX kernel. Performance drops ~40 GFLOPS without it on the M1 in this test.